CancelActivity refactoring

+++ src/org/geometerplus/fbreader/fbreader/FBReaderApp.java
@@ -426,26 +426,12 @@ public final class FBReaderApp extends ZLApplication {
 		}
 	}
 
-	private final ArrayList<CancelMenuHelper.ActionDescription> myCancelActionsList =
-		new ArrayList<CancelMenuHelper.ActionDescription>();
-
 	public List<CancelMenuHelper.ActionDescription> getCancelActionsList() {
-		return getCancelActionsList(Collection);
+		return new CancelMenuHelper().getActionsList(Collection);
 	}
 
-	public List<CancelMenuHelper.ActionDescription> getCancelActionsList(IBookCollection collection) {
-		myCancelActionsList.clear();
-		myCancelActionsList.addAll(new CancelMenuHelper().getActionsList(collection));
-		return myCancelActionsList;
-	}
-
-	public void runCancelAction(int index) {
-		if (index < 0 || index >= myCancelActionsList.size()) {
-			return;
-		}
-
-		final CancelMenuHelper.ActionDescription description = myCancelActionsList.get(index);
-		switch (description.Type) {
+	public void runCancelAction(int index, CancelMenuHelper.ActionType type) {
+		switch (type) {
 			case library:
 				runAction(ActionCode.SHOW_LIBRARY);
 				break;
@@ -457,9 +443,17 @@ public final class FBReaderApp extends ZLApplication {
 				break;
 			case returnTo:
 			{
-				final Bookmark b = ((CancelMenuHelper.BookmarkDescription)description).Bookmark;
-				Collection.deleteBookmark(b);
-				gotoBookmark(b);
+				final List<CancelMenuHelper.ActionDescription> list =
+					new CancelMenuHelper().getActionsList(Collection);
+				if (index < 0 || index >= list.size()) {
+					return;
+				}
+				final CancelMenuHelper.ActionDescription description = list.get(index);
+				if (description instanceof CancelMenuHelper.BookmarkDescription) {
+					final Bookmark b = ((CancelMenuHelper.BookmarkDescription)description).Bookmark;
+					Collection.deleteBookmark(b);
+					gotoBookmark(b);
+				}
 				break;
 			}
 			case close:
